создает компонент рейтинга
