+++ src/index.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 //import { MainComponentProps } from './consts/consts';
 import { App } from './components/app/App';
-import { Offers } from './mock/offers-mock';
+import { offers } from './mock/offers-mock';
 
 const root = ReactDOM.createRoot(
   document.getElementById('root') as HTMLElement
@@ -10,6 +10,6 @@ const root = ReactDOM.createRoot(
 
 root.render(
   <React.StrictMode>
-    <App offers={Offers}/>
+    <App offers={offers}/>
   </React.StrictMode>
 );
