+++ src/mock/offers-mock.ts
@@ -1,10 +1,10 @@
-import { OfferType } from '../consts/consts';
+import { OfferType } from '../types';
 import apartment from './images/apartment.jpg';
 import room from './images/room.jpg';
 import house from './images/house.jpg';
 import apartment2 from './images/apartment2.jpg';
 
-export const Offers: OfferType[] = [
+export const offers: OfferType[] = [
   {
     id: '1',
     title: 'Beautiful & luxurious studio at great location',
@@ -14,19 +14,19 @@ export const Offers: OfferType[] = [
       name: 'Amsterdam',
       location: {
         latitude: 35.35514938496378,
-        longtitude: 44.35514938496378,
+        longitude: 44.35514938496378,
         zoom: 6,
       }
     },
     location: {
       latitude: 35.35514938496378,
-      longtitude: 44.35514938496378,
+      longitude: 44.35514938496378,
       zoom: 6,
     },
     isFavorite: true,
     isPremium: false,
     rating: 5,
-    previewImage: `${apartment}`,
+    previewImage: apartment,
   },
   {
     id: '2',
@@ -37,19 +37,19 @@ export const Offers: OfferType[] = [
       name: 'Brussels',
       location: {
         latitude: 21.35514938496378,
-        longtitude: 99.35514938496378,
+        longitude: 99.35514938496378,
         zoom: 3,
       }
     },
     location: {
       latitude: 21.35514938496378,
-      longtitude: 99.35514938496378,
+      longitude: 99.35514938496378,
       zoom: 3,
     },
     isFavorite: false,
     isPremium: false,
     rating: 4,
-    previewImage: `${room}`,
+    previewImage: room,
   },
   {
     id: '3',
@@ -60,19 +60,19 @@ export const Offers: OfferType[] = [
       name: 'Hamburg',
       location: {
         latitude: 74.35514938496378,
-        longtitude: 55.35514938496378,
+        longitude: 55.35514938496378,
         zoom: 7,
       }
     },
     location: {
       latitude: 74.35514938496378,
-      longtitude: 55.35514938496378,
+      longitude: 55.35514938496378,
       zoom: 7,
     },
     isFavorite: true,
     isPremium: true,
     rating: 5,
-    previewImage: `${house}`,
+    previewImage: house,
   },
   {
     id: '4',
@@ -83,18 +83,18 @@ export const Offers: OfferType[] = [
       name: 'Dusseldorf',
       location: {
         latitude: 65.35514938496378,
-        longtitude: 50.35514938496378,
+        longitude: 50.35514938496378,
         zoom: 2,
       }
     },
     location: {
       latitude: 65.35514938496378,
-      longtitude: 50.35514938496378,
+      longitude: 50.35514938496378,
       zoom: 7,
     },
     isFavorite: true,
     isPremium: false,
     rating: 2,
-    previewImage: `${apartment2}`,
+    previewImage: apartment2,
   },
 ];
